+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 * Improved accuracy of material assignments.
 * Optimized import times for images.
+* Improved the error message when exporting a model with vertex group names not supported by the original wimdo.
 
 ## 0.15.2 - 2024-10-25
 ### Added